refactor(chara-detail): make the factor header flush tolerance pixel-based
## Summary

Express the factor-header flush tolerance in capture pixels instead of a
width-normalized fraction. The two quantities it discriminates -- the ~1 px
header-row detection jitter and the ~2 px scroll at which the content diff
already spikes -- are pixel-scale, not screen-geometry-scale, so a width
fraction drifts with capture resolution for no physical reason: at a larger
capture it admits a bigger scroll, and at a smaller capture it shrinks below the
1 px jitter floor and starts dropping real switches.

## What changed

- `FactorHeaderConfig::flush_tolerance` (0.002, a width fraction) becomes
  `flush_tolerance_px` (1.5 px). At the 736 px live capture this is the same
  ~1.5 px boundary as before, so behaviour there is unchanged.
- `factorHeaderTopY` returns the header top-edge pixel row (`std::optional<int>`)
  instead of a normalized fraction; `reference_header_y` and the gate comparison
  are in pixels. The comparison is valid because both rows are taken on same-size
  frames (already guaranteed by the size check).
- Regenerated `scene_scraper.json`; the reset diagnostic now logs the row in px.

## Testing

- **Golden** `integration_golden`: 100%. **Unit** `umacapture_tests`: 186 cases.
- **Repro clip**: two real-switch resets, both `header_y=12, ref_header_y=12`
  (delta 0), matching the prior fraction-based run (0.0163 * 736 = 12 px).

+++ native/src/chara_detail/chara_detail_scene_scraper.h
@@ -559,11 +559,12 @@ private:
     // re-probes). Reuses the stationary rect and its calibrated color thresholds as the change metric.
     void maybeResetOnFactorChange(const Frame &frame, record::RecordType record_type);
 
-    // Top edge y of the green "因子" section header, as a width-normalized fraction relative to the scroll-area
-    // crop (so it tracks the content, not the scroll thumb). Scans the config band top-down for the first row
-    // that is mostly header green. nullopt when the header is scrolled off or mid-animation (not flush), which
-    // maybeResetOnFactorChange treats as "not at the top". Never throws on a scrolled-away frame.
-    [[nodiscard]] std::optional<double> factorHeaderTopY(const Frame &frame) const;
+    // Top-edge pixel row of the green "因子" section header, relative to the scroll-area crop (so it tracks the
+    // content, not the scroll thumb). Scans the config band top-down for the first row that is mostly header
+    // green. nullopt when the header is scrolled off or mid-animation (not flush), which maybeResetOnFactorChange
+    // treats as "not at the top". Never throws on a scrolled-away frame. Compared against the reference in pixels,
+    // valid because both are taken on same-size frames.
+    [[nodiscard]] std::optional<int> factorHeaderTopY(const Frame &frame) const;
 
     void resetMonitors();
 
@@ -646,9 +647,9 @@ private:
     std::optional<record::RecordType> type_pending_value;
     std::optional<uint64> factor_change_pending_since;
     Frame factor_probe_reference = {};
-    // Header top y (see factorHeaderTopY) captured with factor_probe_reference; the flush gate compares the
-    // current header y against it, so per-device layout differences cancel.
-    std::optional<double> reference_header_y;
+    // Header top-edge pixel row (see factorHeaderTopY) captured with factor_probe_reference; the flush gate
+    // compares the current header row against it in pixels.
+    std::optional<int> reference_header_y;
     std::optional<std::pair<TabPage, bool>> last_scroll_position_emitted;
 };
